requirements.txt
requests==2.31.0
python-dotenv
pandas
psycopg2
google-generativeai